+++ src/lib/auth/cookie.ts
@@ -1,0 +1,7 @@
+import { cookies } from 'next/headers';
+
+export async function getServerSideCookie() {
+  const cookieStore = cookies();
+
+  return cookieStore.get('token')?.value || '';
+}
